Some correction for the helper.c
I help you to correct the errors which is similar to find. 
In my opinion, what you need to update: 1. checking the parameter of each command; 2. the communication in print function (i.e. parent and child talk with each other); 3. print function error checking such as the user is not existed or the scheduler is not existed. 
Good Luck!